+++ src/components/register/TailorRegisterForm.js
